% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_index.R
\name{cor_index}
\alias{cor_index}
\title{Get Correlation Index of DataFrame}
\usage{
cor_index(.df)
}
\arguments{
\item{.df}{object of class \code{data.frame}}
}
\value{
Object of class \code{data.frame} contained correlation index. (See example below.)
}
\description{
Get Correlation Index of DataFrame
}
\examples{
cor_index(mtcars) \%>\% head()
# A tibble: 55 x 3
   Var1  Var2     Cor
   <chr> <chr>  <dbl>
 1 cyl   disp   0.902
 2 disp  wt     0.888
 3 mpg   wt    -0.868
 4 cyl   mpg   -0.852
 5 disp  mpg   -0.848
 6 cyl   hp     0.832
}
