% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLog.R
\name{createlog}
\alias{createlog}
\title{createlog:  Using a single or multiple survey/legs, open the EV file, export the region log, and concatenate 
the logs into a single file called CruiseLog.csv}
\usage{
createlog(SurveyName, DirNameFile)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files}

\item{exportloc}{is an optional argument for using the Export_Dir or the Region_Export_dir column
entry in the DirNameFile spreadsheet.  1 is Export_Dir, and 0 is Region_Export_dir}
}
\description{
Needs RDCOMclient, readxl
}
\examples{
SurveyName='2022_LEG1LOG'
SurveyName=c('2022_LEG1LOG', '2022_LEG2LOG')
DirNameFile='C:/rthomas/R/Rcode/EK60_EK80 conversion/EK60_EK80_conv_updated/Directory Structure EK60 EK80 conversion updated.xlsx'
createlog(SurveyName,DirNameFile)


}
