% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\docType{class}
\name{encoder}
\alias{encoder}
\title{R6 Class For One-Hot-Encoding}
\format{An \code{\link{R6Class}} generator object
TODO add examples}
\usage{
encoder
}
\value{
An object of \code{\link{R6Class}}
}
\description{
R6 Class For One-Hot-Encoding
}
\keyword{datasets}
