% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seasonality.R
\name{extract_seasonality}
\alias{extract_seasonality}
\title{Extract Seasonality}
\usage{
extract_seasonality(...)
}
\value{
data frame with decomposed time series
}
\description{
Extract Seasonality
}
