% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftr_cor_index.R
\name{ftr_cor_index}
\alias{ftr_cor_index}
\title{Get correlation index for data frame}
\usage{
ftr_cor_index(.df)
}
\arguments{
\item{.df}{data frame}
}
\value{
data frame with correllation index
}
\description{
Get correlation index for data frame
}
\examples{
ftr_cor_index(mtcars)

}
