% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_index.R
\name{identity_index}
\alias{identity_index}
\title{Calculate the index of sameness for variables in the data frame}
\usage{
identity_index(.df, .na_include = FALSE)
}
\arguments{
\item{.df}{data frame}

\item{.na_include}{if TRUE, \code{NA == NA} will return \code{TRUE}. 
Otherwise all observations with at least one \code{NA} will go to \code{other} column. 
Default if \code{FALSE}}
}
\value{
data frame with identity index.
  \code{identical} and \code{non_identical} columns contatin number of observations where
  \code{var1 == var2} and \code{var1 != var2} respectively.
  \code{other} column contains number of observations with neither \code{var1 == var2} nor \code{var1 != var2} 
conditions, e.g. \code{NaN}s, \code{NA}s etc.
}
\description{
Calculate the index of sameness for variables in the data frame
}
\examples{
identity_index(mtcars)
}
