% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interactions.R
\name{make_interactions}
\alias{make_interactions}
\title{Make Interactions Over Selected Variables With Given Function}
\usage{
make_interactions(.df, .f, ...)
}
\arguments{
\item{.df}{data frame}

\item{.f}{function}

\item{...}{variables to interact, symbols and tidyselect functions allowed}
}
\description{
For now the problem with naming new columns exists.
It will be solving in the future.
But the main functionality seems to work well.
If you stuck with a bug please create a new issue here 
\url{https://github.com/pavel-filatov/featr/issues}.
}
\examples{
mtcars \%>\% 
  select(am, cyl, mpg) \%>\% 
  mutate(
    !!!make_interactions(., "-", 1, 2),
    !!!make_interactions(., "+", am, cyl),
    !!!make_interactions(., "/", everything())
  )

}
