% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rollmax.R
\name{make_rollmax}
\alias{make_rollmax}
\title{Build a List of Expressions (Rolling Max)}
\usage{
make_rollmax(..., .n = 3, .align = "right", na.rm = FALSE)
}
\arguments{
\item{...}{variables to make \code{roll_max}s}

\item{.n}{integer vector of window sizes to calculate rolling max}

\item{.align}{\code{right}, \code{left} or \code{center}. Default is \code{right}}

\item{na.rm}{are \code{NA}s need to be removed while calculating rolling max? Default is \code{FALSE}}
}
\value{
expression to unquote inside \code{mutate()} function
}
\description{
Build a List of Expressions (Rolling Max)
}
