% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rollmeans.R
\name{make_rollmeans}
\alias{make_rollmeans}
\title{Build a List of Expressions (Rolling Mean)}
\usage{
make_rollmeans(..., .n = 3, .align = "right", na.rm = TRUE)
}
\arguments{
\item{...}{variables to make \code{roll_mean}s}

\item{.n}{integer vector of window sizes to calculate rolling mean}

\item{.align}{\code{right}, \code{left} or \code{center}. Default is \code{right}}

\item{na.rm}{}
}
\value{
expression to unquote inside \code{mutate()} function
}
\description{
Build a List of Expressions (Rolling Mean)
}
