% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shift.R
\name{make_shift}
\alias{make_shift}
\title{Build a Quosure to Make Lags and Leads}
\usage{
make_shift(..., .n = 1)
}
\arguments{
\item{...}{variables}

\item{.n}{integer vector with steps to shift:

    - positive for \code{lag()}  
    
    - negative for \code{lead()}  
    
    - 0 automaticly excluded (so you may use colon notation \code{-2:2})}
}
\value{
expression to unquote inside \code{mutate()} function
}
\description{
This function builds an expression to make multiple lags and/or leads.
\bold{NOTE: For use inside \code{dplyr::mutate()} only!}
}
\details{
Unquotation is needed inside the \code{mutate()} with \code{!!!} (bang-bang-bang) like this:

\code{mutate(mtcars, !!!make_shift(am, mpg))}
}
\examples{
make_shift(am, mpg)
#> $am_lag1
#> dplyr::lag(am, 1)
#> 
#> $mpg_lag1
#> dplyr::lag(mpg, 1)

mutate(mtcars[c("am", "mpg")], !!!make_shift(am, mpg, .n = c(-1, 1, 2)))
#>    am  mpg am_lead1 am_lag1 am_lag2 mpg_lead1 mpg_lag1 mpg_lag2
#> 1   1 21.0        1      NA      NA      21.0       NA       NA
#> 2   1 21.0        1       1      NA      22.8     21.0       NA
#> 3   1 22.8        0       1       1      21.4     21.0     21.0
#> 4   0 21.4        0       1       1      18.7     22.8     21.0
#> 5   0 18.7        0       0       1      18.1     21.4     22.8
#> 6   0 18.1        0       0       0      14.3     18.7     21.4
}
