% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EV_exports.R
\name{read_EV_exports}
\alias{read_EV_exports}
\title{read_EV_exports:  Read in a directory of .csv files exported from Echoview}
\usage{
read_EV_exports(var, SurveyName, DirNameFile)
}
\arguments{
\item{var}{- variables to read in (subdirectories need to be named as variables)}

\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files}
}
\description{
Read in a directory of .csv files (can be of several variables) 
exported from Echoview.   The function preserves the cells by depth. 
Uses packages readxl, readr, purrr, and dplyr.
}
\examples{
temp<-read_EV_exports(variables[1], SurveyName, DirNameFile)

}
