% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EV_exports_v2.R
\name{read_EV_exports_v2}
\alias{read_EV_exports_v2}
\title{read_EV_exports_v2:  Read in a directory of .csv files exported from Echoview}
\usage{
read_EV_exports_v2(SurveyName, DirNameFile, database = 1, exportloc = 1, ...)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files}

\item{database}{is an optional argument for format of the .csv files to.  Default is database (1).  Spreadsheet =0}

\item{exportloc}{is an optional argument for using the Export_Dir or the Region_Export_dir column
entry in the DirNameFile spreadsheet.  1 is Export_Dir, and 0 is Region_Export_dir}
}
\description{
Read in a directory of .csv files
exported from Echoview.   The function preserves the cells by depth. 
Uses packages readxl, readr, purrr, and dplyr.  This version (unlike read_EV_exports) does
not need to have the directory variable
}
\examples{
temp<-read_EV_exports_v2(SurveyName, DirNameFile,database=1,exportloc=1)

}
