% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameator.R
\docType{data}
\name{Renameator}
\alias{Renameator}
\title{Rename all features in data frame}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Renameator
}
\arguments{
\item{df}{input data frame}

\item{pred}{string that will form new variables name,
e.g. "Var1" for `pred = "Var"` or "Question_5" for `pred = "Question_`}

\item{exclude_vars}{which variables should be exclude from renaming process?}
}
\value{
data frame with renamed features
}
\description{
R6 class for renaming dataframes
}
\keyword{datasets}
