% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_with.R
\name{separate_with}
\alias{separate_with}
\title{Separate Data Frame With Any Variable}
\usage{
separate_with(.data, .var)
}
\arguments{
\item{.data}{data frame}

\item{.var}{variable to separate by}
}
\description{
Turn data frame into the list of data frames by unique entries of any variable
}
\examples{
separate_with(mtcars, cyl)
}
