% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetBiomassExpParamsFun.R
\name{SetBiomassExpParamsFun}
\alias{SetBiomassExpParamsFun}
\title{SetBiomassExpParamsFun}
\usage{
SetBiomassExpParamsFun(Obj_ExpVar)
}
\arguments{
\item{Obj_ExpVar-}{an Evfile export object of the form EVfile[['Properties']][['Export']][['Variables']]}
}
\description{
Function to take an EvFile Export object and set the export parameters within it
}
