% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_over.R
\name{sum_over}
\alias{sum_over}
\title{Sum Over Selected Vars}
\usage{
sum_over(.data, ...)
}
\arguments{
\item{.data}{data frame}

\item{...}{variables to sum}
}
\description{
Function builds an expression to sum over selected variables.
You can select variables with tidyverse way.
}
\examples{
# Sum all vars from `mpg` to `carb` except `cyl` to `am`
mutate(mtcars, ss = sum_over(mtcars, mpg:carb, -(cyl:am)))
}
