% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_template_EV.R
\name{update_template_EV}
\alias{update_template_EV}
\title{update_template_EV}
\usage{
update_template_EV(Survey_Name, DirNameFile, ni = 1, ...)
}
\arguments{
\item{DirNameFile}{excel file name of paths}

\item{ni}{is an optional argument for which file to start with (default ni=1 is the first file).

Excel file has the following columns (with header names)
Cal_File: calibration file name
Base_Path:
Orig_EV_Dir:
Raw_Dir
EK80_Cal_File
EK80_Raw_Dir}

\item{SurveyName}{name of survey as in excel file}
}
\description{
function to take in an the survey name and an excel file.  The excel 
file has parameters such as file paths, template paths, etc.
}
\examples{
DirNameFile="C:/rthomas/R/Rcode/EK60_EK80 conversion/EK60_Ek80_conv_updated/Directory Structure EK60 EK80 conversion updated.xlsx"
SurveyName='2019_US'
update_template_EV(SurveyName,DirNameFile, ni=1)
}
