% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featR-package.R
\docType{package}
\name{featR-package}
\alias{featR}
\alias{featR-package}
\title{featR: A Unified Toolkit for Feature Selection}
\description{
Filter, wrapper, and embedded feature-selection methods behind a
consistent set of \verb{fs_*()} functions, plus PCA/SVD dimensionality
reduction helpers. Heavy modeling engines (brms, caret, glmnet, ...) are
optional Suggests; each function checks for what it needs and tells you
what to install.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elkronos/featR}
  \item Report bugs at \url{https://github.com/elkronos/featR/issues}
}

}
\author{
\strong{Maintainer}: Justin Chase \email{jchase.msu@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Justin Chase \email{jchase.msu@gmail.com} [copyright holder]
}

}
\keyword{internal}
